requirements.txt
flask==2.3.2
flask-cors==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
google-generativeai==0.3.1
beautifulsoup4==4.12.2
selenium==4.10.0

# Browser-specific packages
webdriver-manager==3.8.6  # Automatically manages drivers
chromedriver-binary==114.0.5735.90  # For Chrome
msedge-selenium-tools==3.141.3  # For Microsoft Edge
gunicorn==20.1.0  # For production server